Remove the malloc_hook from the pendulum_demo.
glibc 2.34 has removed malloc_hook completely, which means we
can no longer use it (even with deprecations).

That said, it appears that the functionality was broken in
pendulum_demo anyway.  The idea with the malloc_hook was that
it would allow mallocs during initialization, and then disallow
mallocs during execution.  However, I manually added a malloc
during the execution phase, and it did *not* trigger the hook.

Because of all of this, just remove this functionality for now.

Signed-off-by: Chris Lalancette <clalancette@openrobotics.org>