Fix two small bugs in the fibonacci C++ tutorial.
The first bug has to do with the action client, and what happens
if the server rejects the goal.  In the current code, the
client would just hang around forever.  If the server rejects
the goal, make sure the client shuts down.

The second bug has to do with the size that the action server
can support.  The code has more information, but essentially
because of the Fibonacci action message, we can only support
sequences up to order 46; anything larger than that will cause
integer wrapping.

Signed-off-by: Chris Lalancette <clalancette@openrobotics.org>