Dramatically speed up the demo_nodes_cpp tests (#641)
* Don't wait unconditional 5 seconds for processes to quit.

That is, just wait long enough for the process to have
an exit code.  On its own, this change doesn't actually
make a huge difference to the test times, but when combined
with some of the upcoming changes it should make a
significant reduction.

* Minor revamp of the timer tutorials.

Make them use underscores for the member variables,
and make those member variables private.

* Add in the ability to pass parameters to demo_nodes_cpp tests.

That way we can configure some of the demos to be much
more speedy, while retaining their default behavior.

* Add a one-shot feature to add_two_ints_server.

If the one-shot parameter is set, then it will quit
after receiving the first request.  Useful for testing.

* Make content_filtering_publisher have configurable publish time.

This can shorten our tests dramatically.

Signed-off-by: Chris Lalancette <clalancette@gmail.com>