Remove action_tutorials_interfaces. (#701)
We have this message duplicated three times in the core;
once in here, once in example_interfaces, and once in
test_interface_files.  That seems a bit excessive, so
remove this third copy and make the action_tutorials
use the example_interfaces version of Fibonacci.action
instead.

This change obviously changes the dependencies of the
action_tutorials_cpp and action_tutorials_py packages,
in that it removes action_tutorials_interfaces and adds
in example_interfaces.  I believe that this should be
fine from an inter-repository standpoint, as the demos
repository already contains other packages (composition,
quality_of_service_demo_cpp, demo_nodes_py, demo_nodes_cpp)
that depend on example_interfaces.

Signed-off-by: Chris Lalancette <clalancette@gmail.com>